// src/types/PlayerRole.ts
// DO NOT EDIT! THIS IS GENERATED FILE

enum PlayerRole {
  Goalkeeper = 'goalkeeper',
  LeftBack = 'leftBack',
  LeftCenterBack = 'leftCenterBack',
  Centerdef = 'Centerdef',
  RightCenterBack = 'rightCenterBack',
  RightBack = 'rightBack',
  LeftMidfielder = 'leftMidfielder',
  LeftCentralMidfielder = 'leftCentralMidfielder',
  CenterMidfielder = 'centerMidfielder',
  RightCentralMidfielder = 'rightCentralMidfielder',
  RightMidfielder = 'rightMidfielder',
  LeftDefensiveMidfielder = 'leftDefensiveMidfielder',
  LeftCenterDefensiveMidfielder = 'leftCenterDefensiveMidfielder',
  DefensiveMidfielder = 'defensiveMidfielder',
  RightCenterDefensiveMidfielder = 'rightCenterDefensiveMidfielder',
  RightDefensiveMidfielder = 'rightDefensiveMidfielder',
  LeftWinger = 'leftWinger',
  LeftCenterWinger = 'leftCenterWinger',
  CenterForward = 'centerForward',
  RightCenterWinger = 'rightCenterWinger',
  RightWinger = 'rightWinger',
}

export default PlayerRole;